Refactor charts into own cmp/model, small tweaks from CR WRT labeling
