+++ admin/tabs/activity/tracking/charts/ChartsPanel.js
@@ -4,17 +4,18 @@
  *
  * Copyright © 2020 Extremely Heavy Industries Inc.
  */
-
+import {isEmpty} from 'lodash';
 import {uses, hoistCmp} from '@xh/hoist/core';
 import {panel} from '@xh/hoist/desktop/cmp/panel';
 import {tabContainer} from '@xh/hoist/cmp/tab';
 import {toolbar} from '@xh/hoist/desktop/cmp/toolbar';
-import {switchInput} from '../../../../desktop/cmp/input';
-import {ActivityModel} from './ActivityModel';
+import {switchInput} from '@xh/hoist/desktop/cmp/input';
+import {ChartsModel} from './ChartsModel';
 
-export const activityCharts = hoistCmp.factory({
-    model: uses(ActivityModel),
-    render() {
+export const chartsPanel = hoistCmp.factory({
+    model: uses(ChartsModel),
+    render({model}) {
+        if (isEmpty(model.dimensions)) return null;
         return panel({
             item: tabContainer(),
             bbar: bbar(),
